refactor(openomni): extract transcript, use barrel import for compaction
